backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.1.2
